Imported testing framework and refactored express server to be testable
