+++ package.json
@@ -6,7 +6,7 @@
   "license": "UNLICENSED",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "watch": "webpack --watch",
     "dev-server": "nodemon -r dotenv/config server/server.js",
     "dev-client": "webpack-dev-server  --open --color ",
@@ -41,8 +41,10 @@
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "^7.14.2",
     "html-loader": "^0.5.5",
+    "jest": "^24.8.0",
     "prettier": "1.18.2",
     "style-loader": "^0.23.1",
+    "supertest": "^4.0.2",
     "webpack": "^4.35.0",
     "webpack-cli": "^3.3.5",
     "webpack-dev-server": "^3.7.2"
